+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 ## v0.0.6 - 2023-??-?? - ???
 
 * Adds Provider.list_zones to enable new dynamic zone config functionality
+* Newly added support for subnet targeting using the [Subnet traffic-routing method](https://learn.microsoft.com/en-us/azure/traffic-manager/traffic-manager-routing-methods#subnet)
+  in Azure Traffic Managers.
 
 ## v0.0.5 - 2023-07-27 - More feature support